refactor project form
