--- web/src/components/projects/CreateProjectForm.module.css
@@ -1,127 +1,0 @@
-.step-progress {
-  display: flex;
-  justify-content: center;
-  gap: 0.75rem;
-  margin-bottom: 1.5rem;
-}
-
-.step-dot {
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  background-color: #ddd;
-}
-
-.step-dot.active {
-  background-color: var(--ucl-accent);
-  box-shadow: 0 0 0 3px rgba(123, 0, 255, 0.25);
-}
-
-.step-dot.completed {
-  background-color: var(--ucl-mid);
-}
-
-.section-label {
-  font-weight: bold;
-}
-
-.username-input-wrapper {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.username-input-wrapper > div {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.domain-suffix {
-  color: var(--text--secondary);
-  font-weight: 500;
-  font-size: 1rem;
-  white-space: nowrap;
-}
-
-.environment-docs {
-  padding: 0.75rem;
-  background-color: #e7f3ff;
-}
-
-.environment-docs a {
-  color: var(--form-purple--light);
-  text-decoration: none;
-  font-weight: 600;
-}
-
-.environment-docs a:hover {
-  text-decoration: underline;
-}
-
-.role-tags {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.5rem;
-  margin-top: 0.5rem;
-}
-
-.role-tag {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.25rem 0.75rem;
-  background-color: #e7f3ff;
-  border: 1px solid var(--ucl-accent);
-  border-radius: var(--default-border-radius);
-  font-size: 0.875rem;
-  color: var(--ucl-accent);
-}
-
-.role-tooltip span[class*="tooltip__content"] {
-  right: auto;
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.role-tag-remove {
-  background: none;
-  border: none;
-  color: var(--ucl-accent);
-  cursor: pointer;
-  padding: 0;
-  font-size: 1.25rem;
-  line-height: 1;
-  font-weight: bold;
-}
-
-.role-tag-remove:hover {
-  color: var(--ucl-mid);
-}
-
-.role-dropdown {
-  margin-top: 0.5rem;
-}
-
-.researcher-content {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.researcher-content > div {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.actions {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid var(--color-border);
-}
