--- web/src/components/projects/CreateProjectForm.tsx
@@ -1,732 +1,0 @@
-import { useState, useEffect } from "react";
-import { Controller, useFieldArray, useForm } from "react-hook-form";
-import {
-  postProjectsTre,
-  putProjectsTreByProjectId,
-  ProjectTreRequest,
-  ProjectTreRoleName,
-  Study,
-  Asset,
-  Environment,
-  getStudiesByStudyIdAssets,
-  getEnvironments,
-} from "@/openapi";
-import { AnyProject, AnyProjectRoleName } from "@/types/projects";
-import { extractErrorMessage, responseIsError } from "@/lib/errorHandler";
-import Button from "../ui/Button";
-import Dialog from "../ui/Dialog";
-import InfoTooltip from "../ui/InfoTooltip";
-import { HelperText, Alert, AlertMessage } from "../shared/uikitExports";
-import { ROLE_LABELS, ROLE_DESCRIPTIONS, getAvailableRoles, getProjectNameValidation } from "./lib/projects";
-
-import styles from "./CreateProjectForm.module.css";
-
-// this should match the domain that is used for the entra ID users in the portal
-const domainName = process.env.NEXT_PUBLIC_DOMAIN_NAME || "@ucl.ac.uk";
-
-type ProjectFormData = {
-  name: string;
-  studyId: string;
-  environmentId: string;
-  assetIds: { value: string }[];
-  additionalApprovedResearchers: { username: string; roles: AnyProjectRoleName[] }[];
-};
-
-type Props = {
-  approvedStudies: Study[];
-  handleProjectCreated: () => void;
-  handleCancelCreate: () => void;
-  editingProject?: AnyProject | null;
-};
-
-export default function CreateProjectForm({
-  approvedStudies,
-  handleProjectCreated,
-  handleCancelCreate,
-  editingProject,
-}: Props) {
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [assets, setAssets] = useState<Asset[]>([]);
-  const [isLoadingAssets, setIsLoadingAssets] = useState(false);
-  const [environments, setEnvironments] = useState<Environment[]>([]);
-  const [isLoadingEnvironments, setIsLoadingEnvironments] = useState(true);
-  const [environmentsError, setEnvironmentsError] = useState<string | null>(null);
-  const [currentStep, setCurrentStep] = useState(1);
-  const [originalMemberCount, setOriginalMemberCount] = useState(0);
-  const totalSteps = 2;
-
-  const {
-    register,
-    handleSubmit,
-    watch,
-    setValue,
-    control,
-    getValues,
-    trigger,
-    formState: { errors },
-  } = useForm<ProjectFormData>({
-    mode: "onChange",
-    defaultValues: {
-      name: "",
-      studyId: "",
-      environmentId: "",
-      assetIds: [],
-      additionalApprovedResearchers: [],
-    },
-  });
-
-  const selectedStudyId = watch("studyId");
-  const selectedEnvironmentId = watch("environmentId");
-  const selectedAssetIds = watch("assetIds");
-  const {
-    fields: assetFields,
-    append: appendAsset,
-    remove: removeAsset,
-  } = useFieldArray({
-    control,
-    name: "assetIds",
-  });
-
-  const {
-    fields: researcherFields,
-    append: appendResearcher,
-    remove: removeResearcher,
-  } = useFieldArray({
-    control,
-    name: "additionalApprovedResearchers",
-  });
-
-  // Populate form with existing project data when editing
-  useEffect(() => {
-    if (editingProject) {
-      setValue("name", editingProject.name);
-      setValue("studyId", editingProject.study_id);
-
-      const environment = environments.find((env) => env.name === editingProject.environment_name);
-      if (environment) {
-        setValue("environmentId", environment.id);
-      }
-
-      const projectAssetIds = editingProject.assets?.map((asset) => ({ value: asset.id })) || [];
-      setValue("assetIds", projectAssetIds);
-      setAssets(editingProject.assets || []);
-
-      const projectMembers =
-        editingProject.members?.map((member) => ({
-          username: member.username.replace(domainName, ""),
-          roles: member.roles as AnyProjectRoleName[],
-        })) || [];
-      setValue("additionalApprovedResearchers", projectMembers);
-      setOriginalMemberCount(projectMembers.length);
-    }
-  }, [editingProject, setValue, environments]);
-
-  useEffect(() => {
-    if (!editingProject) {
-      setValue("assetIds", []);
-    }
-  }, [selectedEnvironmentId, setValue, editingProject]);
-
-  useEffect(() => {
-    const fetchEnvironments = async () => {
-      setIsLoadingEnvironments(true);
-      setEnvironmentsError(null);
-      try {
-        const response = await getEnvironments();
-        if (responseIsError(response) || !response.data) {
-          const errorMsg = extractErrorMessage(response);
-          setEnvironmentsError(`Failed to load environments: ${errorMsg}`);
-          return;
-        }
-        setEnvironments(response.data);
-      } catch (error) {
-        console.error("Failed to fetch environments:", error);
-        setEnvironmentsError("Failed to load environments. Please try again later.");
-      } finally {
-        setIsLoadingEnvironments(false);
-      }
-    };
-
-    fetchEnvironments();
-  }, []);
-
-  // fetch assets dynamically based on selected study
-  useEffect(() => {
-    const fetchAssets = async () => {
-      if (!selectedStudyId || !selectedEnvironmentId) {
-        setAssets([]);
-        return;
-      }
-
-      setIsLoadingAssets(true);
-      try {
-        const response = await getStudiesByStudyIdAssets({
-          path: { studyId: selectedStudyId },
-        });
-
-        if (responseIsError(response) || !response.data) {
-          const errorMsg = extractErrorMessage(response);
-          setError(`Failed to fetch assets: ${errorMsg}`);
-          setAssets([]);
-          return;
-        }
-        setAssets(response.data);
-      } catch (error) {
-        console.error("Failed to fetch assets:", error);
-        setError("Failed to fetch assets. Please try again.");
-        setAssets([]);
-      } finally {
-        setIsLoadingAssets(false);
-      }
-    };
-
-    fetchAssets();
-  }, [selectedStudyId, selectedEnvironmentId]);
-
-  const nextStep = async () => {
-    if (currentStep === totalSteps) return;
-
-    // Validate step 1 fields before proceeding
-    if (currentStep === 1) {
-      const isValid = await trigger(["name", "studyId", "environmentId"]);
-      if (!isValid) return;
-    }
-
-    setCurrentStep(currentStep + 1);
-  };
-
-  const prevStep = () => {
-    if (currentStep !== 1) setCurrentStep(currentStep - 1);
-  };
-
-  const submitProject = async (data: ProjectFormData) => {
-    setIsSubmitting(true);
-    setError(null);
-
-    try {
-      // Find the selected environment to determine which endpoint to call
-      const selectedEnvironment = environments.find((env) => env.id === data.environmentId);
-      if (!selectedEnvironment) throw new Error("Selected environment not found");
-
-      const assetIds = data.assetIds.map((asset) => asset.value).filter((id) => id !== "");
-
-      let response;
-
-      switch (selectedEnvironment.name) {
-        case "ARC Trusted Research Environment":
-          const treMembers = data.additionalApprovedResearchers
-            .filter((researcher) => researcher.username.trim() !== "")
-            .map((researcher) => ({
-              username: `${researcher.username.trim()}${domainName}`,
-              roles: researcher.roles as ProjectTreRoleName[],
-            }));
-
-          if (editingProject) {
-            // Update existing project (only members and assets)
-            response = await putProjectsTreByProjectId({
-              path: { projectId: editingProject.id },
-              body: {
-                asset_ids: assetIds,
-                members: treMembers,
-              },
-            });
-          } else {
-            // Create new project
-            const requestBody: ProjectTreRequest = {
-              name: data.name,
-              study_id: data.studyId,
-              asset_ids: assetIds,
-              members: treMembers,
-            };
-            response = await postProjectsTre({ body: requestBody });
-          }
-          break;
-        case "Data Safe Haven":
-          throw new Error("DSH projects are not yet supported");
-        default:
-          throw new Error(`Unknown environment: ${selectedEnvironment.name}`);
-      }
-
-      if (responseIsError(response)) {
-        const errorMsg = extractErrorMessage(response);
-        setError(errorMsg);
-        return;
-      }
-
-      handleProjectCreated();
-    } catch (error) {
-      console.error(error);
-      setError("Something went wrong. Failed to create or update project.");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  return (
-    <Dialog setDialogOpen={handleCancelCreate}>
-      <div className={styles.container}>
-        <h2>{editingProject ? "Edit Project" : "Create New Project"}</h2>
-
-        <div className={styles["step-progress"]}>
-          {Array.from({ length: totalSteps }, (_, i) => i + 1).map((step) => (
-            <div
-              key={step}
-              className={`${styles["step-dot"]} ${currentStep === step ? styles["active"] : ""} ${
-                currentStep > step ? styles["completed"] : ""
-              }`}
-            />
-          ))}
-        </div>
-
-        {error && (
-          <Alert type="error">
-            <AlertMessage>{error}</AlertMessage>
-          </Alert>
-        )}
-
-        <form className="form" onSubmit={(e) => e.preventDefault()}>
-          {currentStep === 1 && (
-            <>
-              <div className="field">
-                <label htmlFor="studyId">Study *</label>
-                <select
-                  id="studyId"
-                  className={styles.select}
-                  {...register("studyId", {
-                    required: "Please select a study",
-                  })}
-                  disabled={isSubmitting || !!editingProject}
-                >
-                  <option value="">Select a study...</option>
-                  {approvedStudies.map((approvedStudy) => (
-                    <option key={approvedStudy.id} value={approvedStudy.id}>
-                      {approvedStudy.title}
-                    </option>
-                  ))}
-                </select>
-                {errors.studyId && (
-                  <Alert type="error">
-                    <AlertMessage>{errors.studyId.message}</AlertMessage>
-                  </Alert>
-                )}
-                <HelperText>Select the study this project will belong to</HelperText>
-              </div>
-
-              <div className="field">
-                <label htmlFor="environmentId">Environment *</label>
-                <select
-                  id="environmentId"
-                  className={styles.select}
-                  {...register("environmentId", {
-                    required: "Please select an environment",
-                  })}
-                  disabled={isSubmitting || isLoadingEnvironments || !!environmentsError || !!editingProject}
-                >
-                  <option value="">
-                    {isLoadingEnvironments
-                      ? "Loading environments..."
-                      : environmentsError
-                        ? "Error loading environments"
-                        : "Select an environment..."}
-                  </option>
-                  {/* TODO: Remove this filter once DSH project creation is supported */}
-                  {environments
-                    .filter((environment) => environment.name !== "Data Safe Haven")
-                    .map((environment) => (
-                      <option key={environment.id} value={environment.id}>
-                        {environment.name} (Tier {environment.tier})
-                      </option>
-                    ))}
-                </select>
-                {environmentsError && (
-                  <Alert type="error">
-                    <AlertMessage>{environmentsError}</AlertMessage>
-                  </Alert>
-                )}
-                {errors.environmentId && (
-                  <Alert type="error">
-                    <AlertMessage>{errors.environmentId.message}</AlertMessage>
-                  </Alert>
-                )}
-                <HelperText>Select the environment where this project will be deployed</HelperText>
-
-                {selectedEnvironmentId &&
-                  (() => {
-                    const selectedEnvironment = environments.find((env) => env.id === selectedEnvironmentId);
-                    if (!selectedEnvironment) return null;
-
-                    switch (selectedEnvironment.name) {
-                      case "ARC Trusted Research Environment":
-                        return (
-                          <div className={styles["environment-docs"]}>
-                            <a href="https://docs.tre.arc.ucl.ac.uk/" target="_blank" rel="noopener noreferrer">
-                              View TRE documentation
-                            </a>
-                          </div>
-                        );
-                      case "Data Safe Haven":
-                        return (
-                          <div className={styles["environment-docs"]}>
-                            <a
-                              href="https://www.ucl.ac.uk/isd/services/file-storage-sharing/data-safe-haven-dsh"
-                              target="_blank"
-                              rel="noopener noreferrer"
-                            >
-                              View DSH documentation
-                            </a>
-                          </div>
-                        );
-                      default:
-                        return null;
-                    }
-                  })()}
-              </div>
-
-              <div className="field">
-                <label htmlFor="name">Project Name *</label>
-                <Controller
-                  name="name"
-                  control={control}
-                  rules={{
-                    required: "Project name is required",
-                    validate: {
-                      format: (value) => {
-                        const selectedEnvironment = environments.find((env) => env.id === selectedEnvironmentId);
-                        if (!selectedEnvironment) return true; // Skip validation if no environment selected
-
-                        const validation = getProjectNameValidation(selectedEnvironment.name);
-
-                        if (value.length < validation.minLength) {
-                          return `Project name must be at least ${validation.minLength} characters`;
-                        }
-                        if (value.length > validation.maxLength) {
-                          return `Project name must be less than ${validation.maxLength} characters`;
-                        }
-                        if (!validation.pattern.test(value)) {
-                          return validation.patternMessage;
-                        }
-                        return true;
-                      },
-                    },
-                  }}
-                  render={({ field }) => (
-                    <input
-                      {...field}
-                      id="name"
-                      type="text"
-                      placeholder="e.g., myproject"
-                      disabled={
-                        isSubmitting ||
-                        !!editingProject ||
-                        !environments.find((env) => env.id === selectedEnvironmentId)
-                      }
-                    />
-                  )}
-                />
-
-                {errors.name && (
-                  <Alert type="error">
-                    <AlertMessage>{errors.name.message}</AlertMessage>
-                  </Alert>
-                )}
-                <HelperText>
-                  {(() => {
-                    const selectedEnvironment = environments.find((env) => env.id === selectedEnvironmentId);
-                    if (!selectedEnvironment) return "Select an environment to see naming requirements";
-                    return getProjectNameValidation(selectedEnvironment.name).helperText;
-                  })()}
-                </HelperText>
-              </div>
-
-              <div className="field">
-                <span className={styles["section-label"]}>Add assets (optional):</span>
-                <fieldset className="linkage-fieldset">
-                  {assetFields.map((field, index) => (
-                    <div key={field.id} className="item-wrapper">
-                      <label htmlFor={`asset-${index}`} className="item-label">
-                        Asset {index + 1}:
-                      </label>
-                      <Controller
-                        name={`assetIds.${index}.value` as const}
-                        control={control}
-                        render={({ field }) => (
-                          <select
-                            {...field}
-                            id={`asset-${index}`}
-                            className={styles.select}
-                            disabled={isSubmitting || !selectedStudyId || !selectedEnvironmentId || isLoadingAssets}
-                          >
-                            <option value="">
-                              {!selectedStudyId
-                                ? "Select a study first..."
-                                : !selectedEnvironmentId
-                                  ? "Select an environment first..."
-                                  : isLoadingAssets
-                                    ? "Loading assets..."
-                                    : assets.length === 0
-                                      ? "No assets available for this study"
-                                      : "Select an asset (optional)..."}
-                            </option>
-                            {assets.map((asset) => {
-                              const selectedEnvironment = environments.find((env) => env.id === selectedEnvironmentId);
-                              const isCompatible =
-                                !selectedEnvironmentId ||
-                                !selectedEnvironment ||
-                                asset.tier <= selectedEnvironment.tier;
-                              const isAlreadySelected = selectedAssetIds.some(
-                                (selected, selectedIndex) => selected.value === asset.id && selectedIndex !== index
-                              );
-                              const label = isCompatible
-                                ? `${asset.title} (Tier ${asset.tier})`
-                                : `${asset.title} (Tier ${asset.tier}) - Incompatible with selected environment (max tier ${selectedEnvironment?.tier})`;
-
-                              return (
-                                <option key={asset.id} value={asset.id} disabled={!isCompatible || isAlreadySelected}>
-                                  {label}
-                                  {isAlreadySelected ? " - Already selected" : ""}
-                                </option>
-                              );
-                            })}
-                          </select>
-                        )}
-                      />
-                      <Button
-                        onClick={() => removeAsset(index)}
-                        className="remove-button"
-                        aria-label={`Remove contract ${index + 1}`}
-                        type="button"
-                      >
-                        ×
-                      </Button>
-                    </div>
-                  ))}
-                  <Button
-                    onClick={() => appendAsset({ value: "" })}
-                    type="button"
-                    variant="secondary"
-                    size="small"
-                    aria-label="Add Asset"
-                  >
-                    Add Asset
-                  </Button>
-                </fieldset>
-                <HelperText>
-                  Optionally link this project to one or more existing assets from the selected study
-                </HelperText>
-              </div>
-            </>
-          )}
-
-          {currentStep === 2 && (
-            <>
-              <div className="field">
-                <span className={styles["section-label"]}>Project users (optional):</span>
-                <fieldset className="linkage-fieldset">
-                  {researcherFields.map((field, index) => {
-                    const researcher = watch(`additionalApprovedResearchers.${index}`);
-                    const selectedEnvironment = environments.find((env) => env.id === selectedEnvironmentId);
-                    const availableRoles = selectedEnvironment ? getAvailableRoles(selectedEnvironment.name) : [];
-                    const availableRolesToAdd = availableRoles.filter((role) => !researcher?.roles?.includes(role));
-
-                    return (
-                      <div key={field.id} className="item-wrapper">
-                        <div className={styles["researcher-content"]}>
-                          <div>
-                            <label htmlFor={`researcher-${index}`} className={styles["field-label"]}>
-                              Username:
-                            </label>
-                            <Controller
-                              name={`additionalApprovedResearchers.${index}.username` as const}
-                              control={control}
-                              rules={{
-                                required: "Username is required",
-                                validate: {
-                                  isNotEmpty: (value) => {
-                                    if (!value || value.trim() === "") {
-                                      return "Username is required";
-                                    }
-                                    return true;
-                                  },
-                                  notEmailPart: (value) => {
-                                    if (value.includes("@")) {
-                                      return `Enter only the username part (without ${domainName})`;
-                                    }
-                                    return true;
-                                  },
-                                  isUnique: (value) => {
-                                    const allUsernames = getValues("additionalApprovedResearchers").map(
-                                      (user) => user.username
-                                    );
-                                    const duplicateCount = allUsernames.filter((username) => username === value).length;
-                                    return duplicateCount <= 1 || "Username has already been entered";
-                                  },
-                                },
-                              }}
-                              render={({ field: usernameField, fieldState }) => (
-                                <div className={styles["username-input-wrapper"]}>
-                                  <div>
-                                    <input
-                                      {...usernameField}
-                                      id={`researcher-${index}`}
-                                      type="text"
-                                      placeholder="ccaxyz"
-                                      className={styles.select}
-                                      disabled={isSubmitting || (!!editingProject && index < originalMemberCount)}
-                                    />
-                                    <span className={styles["domain-suffix"]}>{domainName}</span>
-                                  </div>
-                                  {fieldState.error && (
-                                    <Alert type="error">
-                                      <AlertMessage>{fieldState.error.message}</AlertMessage>
-                                    </Alert>
-                                  )}
-                                </div>
-                              )}
-                            />
-                          </div>
-
-                          <div>
-                            <label className={styles["field-label"]}>Roles:</label>
-                            <Controller
-                              name={`additionalApprovedResearchers.${index}.roles` as const}
-                              control={control}
-                              rules={{
-                                validate: {
-                                  hasAtLeastOneRole: (value) => {
-                                    if (!value || value.length === 0) {
-                                      return "At least one role is required";
-                                    }
-                                    return true;
-                                  },
-                                },
-                              }}
-                              render={({ fieldState }) => (
-                                <div>
-                                  {researcher?.roles && researcher.roles.length > 0 && (
-                                    <div className={styles["role-tags"]}>
-                                      {researcher.roles.map((role) => (
-                                        <span key={role} className={styles["role-tag"]}>
-                                          {ROLE_LABELS[role]}
-                                          <span className={styles["role-tooltip"]}>
-                                            <InfoTooltip text={ROLE_DESCRIPTIONS[role]} />
-                                          </span>
-                                          <button
-                                            type="button"
-                                            className={styles["role-tag-remove"]}
-                                            onClick={() => {
-                                              const currentRoles = getValues(
-                                                `additionalApprovedResearchers.${index}.roles`
-                                              );
-                                              setValue(
-                                                `additionalApprovedResearchers.${index}.roles`,
-                                                currentRoles.filter((chosenRole) => chosenRole !== role),
-                                                { shouldValidate: true }
-                                              );
-                                            }}
-                                            aria-label={`Remove ${role} role`}
-                                          >
-                                            ×
-                                          </button>
-                                        </span>
-                                      ))}
-                                    </div>
-                                  )}
-                                  {availableRolesToAdd.length > 0 && (
-                                    <select
-                                      className={`${styles.select} ${styles["role-dropdown"]}`}
-                                      value=""
-                                      onChange={(e) => {
-                                        if (e.target.value) {
-                                          const currentRoles =
-                                            getValues(`additionalApprovedResearchers.${index}.roles`) || [];
-                                          setValue(
-                                            `additionalApprovedResearchers.${index}.roles`,
-                                            [...currentRoles, e.target.value as AnyProjectRoleName],
-                                            { shouldValidate: true }
-                                          );
-                                        }
-                                      }}
-                                      disabled={isSubmitting}
-                                    >
-                                      <option value="">+ Add role...</option>
-                                      {availableRolesToAdd.map((role) => (
-                                        <option key={role} value={role}>
-                                          {ROLE_LABELS[role]}
-                                        </option>
-                                      ))}
-                                    </select>
-                                  )}
-                                  {fieldState.error && (
-                                    <Alert type="error">
-                                      <AlertMessage>{fieldState.error.message}</AlertMessage>
-                                    </Alert>
-                                  )}
-                                </div>
-                              )}
-                            />
-                          </div>
-                        </div>
-
-                        <button
-                          type="button"
-                          onClick={() => removeResearcher(index)}
-                          className="remove-button"
-                          aria-label={`Remove researcher ${index + 1}`}
-                        >
-                          ×
-                        </button>
-                      </div>
-                    );
-                  })}
-
-                  <Button
-                    className={styles["add-button"]}
-                    type="button"
-                    variant="secondary"
-                    size="small"
-                    onClick={() => appendResearcher({ username: "", roles: [] })}
-                  >
-                    Add Researcher
-                  </Button>
-                </fieldset>
-                <HelperText>Optionally add additional researchers with their roles to this project</HelperText>
-              </div>
-            </>
-          )}
-
-          <div className={styles.actions}>
-            {currentStep > 1 && (
-              <Button type="button" variant="secondary" onClick={prevStep}>
-                ← Back
-              </Button>
-            )}
-
-            {currentStep < totalSteps && (
-              <Button type="button" onClick={nextStep} cy="next-form-page-button">
-                Next →
-              </Button>
-            )}
-
-            {currentStep === totalSteps && (
-              <Button
-                type="button"
-                disabled={isSubmitting}
-                onClick={handleSubmit((data) => submitProject(data))}
-                cy="submit-project-button"
-              >
-                {isSubmitting
-                  ? editingProject
-                    ? "Updating..."
-                    : "Creating..."
-                  : editingProject
-                    ? "Update Project"
-                    : "Save Draft"}
-              </Button>
-            )}
-          </div>
-        </form>
-      </div>
-    </Dialog>
-  );
-}
